Update for nodiscard correctness in core

+++ 3ds/source/picoc/pksm_api.cpp
@@ -2650,7 +2650,8 @@ void pksm_get_max_pp(
 
     if (move <= int(pksm::Move::SurgingStrikes))
     {
-        ReturnValue->Val->Integer = pksm::VersionTables::movePP(gen, pksm::Move{move}, ups);
+        ReturnValue->Val->Integer =
+            pksm::VersionTables::movePP(gen, pksm::Move{static_cast<u16>(move)}, ups);
     }
     else
     {
@@ -2661,22 +2662,22 @@ void pksm_get_max_pp(
 void pksm_bz2_decompress(
     struct ParseState* Parser, struct Value* ReturnValue, struct Value** Param, int NumArgs)
 {
-    u8** out = (u8**)Param[0]->Val->Pointer;
+    u8** out     = (u8**)Param[0]->Val->Pointer;
     int* outSize = (int*)Param[1]->Val->Pointer;
-    u8* data = (u8*)Param[2]->Val->Pointer;
-    int size = Param[3]->Val->Integer;
+    u8* data     = (u8*)Param[2]->Val->Pointer;
+    int size     = Param[3]->Val->Integer;
 
     std::vector<u8> outData;
     if (BZ2::decompress(data, size, outData) != BZ_OK)
     {
         ReturnValue->Val->Integer = 0;
-        *out = nullptr;
-        *outSize = 0;
+        *out                      = nullptr;
+        *outSize                  = 0;
     }
     else
     {
         ReturnValue->Val->Integer = 1;
-        *out = (u8*) malloc(outData.size());
+        *out                      = (u8*)malloc(outData.size());
         std::copy(outData.begin(), outData.end(), *out);
         *outSize = outData.size();
     }
@@ -2685,22 +2686,22 @@ void pksm_bz2_decompress(
 void pksm_bz2_compress(
     struct ParseState* Parser, struct Value* ReturnValue, struct Value** Param, int NumArgs)
 {
-    u8** out = (u8**)Param[0]->Val->Pointer;
+    u8** out     = (u8**)Param[0]->Val->Pointer;
     int* outSize = (int*)Param[1]->Val->Pointer;
-    u8* data = (u8*)Param[2]->Val->Pointer;
-    int size = Param[3]->Val->Integer;
+    u8* data     = (u8*)Param[2]->Val->Pointer;
+    int size     = Param[3]->Val->Integer;
 
     std::vector<u8> outData;
     if (BZ2::compress(outData, data, size) != BZ_OK)
     {
         ReturnValue->Val->Integer = 0;
-        *out = nullptr;
-        *outSize = 0;
+        *out                      = nullptr;
+        *outSize                  = 0;
     }
     else
     {
         ReturnValue->Val->Integer = 1;
-        *out = (u8*)malloc(outData.size());
+        *out                      = (u8*)malloc(outData.size());
         std::copy(outData.begin(), outData.end(), *out);
         *outSize = outData.size();
     }
